src/requirements.txt
git+https://github.com/idealo/image-super-resolution#egg=ISR[gpu]
scikit-image
boto3==1.9.216
requests
flask
opencv-python
Pillow
patool
pyunpack
numpy